Fixes `__eq__` and `__ne__` operators to accept `object`
The `SourceLocation`, `SourceRange`, and `Cursor` `__eq__` and `__ne__`
functions originally only accepted their self types as second arguments.
This was due to the Clang binding implementation throwing when any other
type was passed in. However, the annotation violates Liskov Substitution
Principle, which makes mypy angry...so apparently it's better to just
accept `object` here and fix the Clang bindings in the future.